+++ lib/kommander-buildah/src/commonMain/kotlin/domain/util/LogLevel.kt
@@ -1,0 +1,6 @@
+package dev.petuska.kommander.buildah.domain.util
+
+@Suppress("EnumEntryName")
+public enum class LogLevel {
+  trace, debug, info, warn, error, fatal, panic
+}