+++ lib/kommander-core/src/commonMain/kotlin/Command.kt
@@ -1,0 +1,6 @@
+package dev.petuska.kommander.core
+
+import dev.petuska.kommander.core.util.Named
+
+public typealias GenericCommand = Command<*>
+public interface Command<O : Option> : Named