+++ lib/kommander-core/src/commonMain/kotlin/Option.kt
@@ -1,0 +1,5 @@
+package dev.petuska.kommander.core
+
+import dev.petuska.kommander.core.util.NamedAndValued
+
+public interface Option : NamedAndValued<String?>