+++ lib/kommander-core/src/commonMain/kotlin/util/executeCommand.kt
@@ -1,0 +1,5 @@
+package dev.petuska.kommander.core.util
+
+import dev.petuska.kommander.core.CMD
+
+internal expect fun executeCommand(cmd: CMD, stdout: (String) -> Unit): Int