+++ lib/kommander-core/src/commonTest/kotlin/PlaceholderTest.kt
@@ -1,0 +1,11 @@
+package dev.petuska.kommander.core
+
+import local.test.BlockingTest
+import kotlin.test.Test
+
+class PlaceholderTest : BlockingTest {
+  @Test
+  fun test() = blockingTest {
+    println("Nothing Yet!")
+  }
+}