+++ lib/kommander-podman/src/commonTest/kotlin/PodmanTest.kt
@@ -1,0 +1,11 @@
+package dev.petuska.kommander.podman
+
+import local.test.BlockingTest
+import kotlin.test.Test
+
+class PodmanTest : BlockingTest {
+  @Test
+  fun test() = blockingTest {
+    podman()
+  }
+}