+++ lib/kontainers-buildah/src/commonMain/kotlin/domain/util/LogLevel.kt
@@ -1,6 +1,0 @@
-package dev.petuska.kontainers.buildah.domain.util
-
-@Suppress("EnumEntryName")
-public enum class LogLevel {
-  trace, debug, info, warn, error, fatal, panic
-}