+++ lib/kontainers-core/src/commonMain/kotlin/Command.kt
@@ -1,6 +1,0 @@
-package dev.petuska.kontainers.core
-
-import dev.petuska.kontainers.core.util.Named
-
-public typealias GenericCommand = Command<*>
-public interface Command<O : Option> : Named