+++ lib/kontainers-core/src/commonMain/kotlin/Option.kt
@@ -1,5 +1,0 @@
-package dev.petuska.kontainers.core
-
-import dev.petuska.kontainers.core.util.NamedAndValued
-
-public interface Option : NamedAndValued<String?>