+++ lib/kontainers-core/src/commonMain/kotlin/util/executeCommand.kt
@@ -1,3 +1,0 @@
-package dev.petuska.kontainers.core.util
-
-internal expect fun executeCommand(cmd: String, onEachLine: (String) -> Unit): Int