+++ lib/kontainers-podman/src/commonMain/kotlin/index.kt
@@ -1,4 +1,0 @@
-package dev.petuska.kontainers.podman
-
-public fun podman() {
-}