+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v0.0.1
 ## Changes
-* N\A
+* kommander-core module
+* commander-buildah module